refactor: split CriandoUmClube in another component
